refactor: set order domain in common package because not split yet
