+++ src/main/java/kitchenpos/common/domain/OrderLineItem.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.common.domain;
+
+import kitchenpos.menu.domain.Menu;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
