+++ src/main/java/kitchenpos/common/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.common.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
